refactor(cli): share agent model argument parsing
